refactor(answer): give get_answer's helpers their own modules
answer.py had grown to 2,805 lines and was doing every job in the tool at
once: sequencing retrieval, grading the result, building six different reply
shapes, deciding whether a cached row could be served, and holding every
feature flag. Splitting a function inside the file makes it longer, not
shorter, so the reduction comes from moving code out to modules that own one
decision each.

Pure moves plus comment edits. No logic, ordering, condition, threshold or
payload key changes, and _ANSWER_SCHEMA_VERSION is untouched.

_run_retrieval_pipeline deliberately stays. It looked like the obvious
candidate for retrieval.py, but the unit suite patches _hybrid_retrieve,
_hydrate_hits and _anchor_symbol_hits on this module; moving the pipeline
would detach those patches and leave the tests green while exercising the
real retriever.

The comments lost their internal corpus evidence — tree names, measurement
dates, sample sizes, byte and millisecond figures — and kept every reason,
trap and rejected alternative.
