+++ packages/server/src/repowise/server/mcp_server/tool_answer/confidence.py
@@ -1,16 +1,40 @@
-"""Confidence-signal helpers for get_answer.
+"""Confidence and retrieval_quality grading for get_answer.
 
-The confidence / retrieval_quality gating is sequenced inline by the
-orchestrator (it is interleaved with payload construction), but the hedge
-detector — the gate that overrides dominance when the LLM admits it can't
-answer — is a pure predicate and lives here.
+Two ratings that answer different questions. ``confidence`` says how much to
+trust the synthesised text; ``retrieval_quality`` says how good the retrieval
+that fed it was. The agent reads the first to decide whether to re-read the
+source, the second to decide whether to search again.
+
+:func:`_grade_answer` runs the gate cascade — one starting grade from retrieval
+dominance, then a run of gates that can only demote it — and the predicates each
+gate reads live beside it.
 """
 
 from __future__ import annotations
 
 import re
+from typing import NamedTuple
 
-from repowise.server.mcp_server.tool_answer.config import _HEDGE_MARKERS
+from repowise.server.mcp_server._answer_context import (
+    is_mechanism_question as _is_mechanism_question,
+)
+from repowise.server.mcp_server._answer_context import (
+    is_why_question as _is_why_question,
+)
+from repowise.server.mcp_server.tool_answer.config import (
+    _AGREEMENT_RANK_GAP,
+    _AGREEMENT_TOP_RANK_MAX,
+    _CLAIM_SUPPORT_GATE_ENV,
+    _DOMINANCE_RATIO,
+    _EARN_HIGH_GROUNDING_ENV,
+    _ENRICH_TOP_N_HITS,
+    _HEDGE_MARKERS,
+    _HIGH_CONFIDENCE_SCORE_FLOOR,
+    _INLINE_BODY_MAX_LINES,
+    _SYMBOL_AGREEMENT_TOP_RANK_MAX,
+    _flag_on,
+)
+from repowise.server.mcp_server.tool_answer.symbols import is_symbol_lookup_question
 
 
 def _answer_is_hedged(answer_text: str) -> bool:
@@ -387,3 +411,389 @@ def implicated_withheld_symbols(
             seen.add(name)
             out.append(name)
     return out
+
+
+def _top_two_score_ratio(hits: list[dict]) -> float:
+    """The top hit's retrieval score over the runner-up's.
+
+    The separator both the confidence grade and :func:`_retrieval_quality` are
+    built on, kept in one place because they must agree on what "dominant"
+    measures. A lone hit has nothing to be ambiguous against, so it is infinitely
+    dominant; no hits at all is zero.
+    """
+    if len(hits) >= 2:
+        return hits[0].get("score", 0.0) / (hits[1].get("score", 0.0) or 1e-9)
+    return float("inf") if hits else 0.0
+
+
+def _agreement_dominant(hits: list[dict], *, vector_leg_keyless: bool = False) -> bool:
+    """True when the top hit is the confident pick by retriever AGREEMENT.
+
+    RRF fusion compresses scores: a page both retrievers rank #1 barely
+    outscores one they rank #2, so the numeric dominance ratio calls the *most*
+    confident retrieval "non-dominant" and demotes it. This reads the per-source
+    ranks instead: when two retrievers put the SAME page at (or within a rank of)
+    the top, that consensus is a stronger ground-truth signal than any RRF score
+    margin.
+
+    Conservative. Requires the top hit to be found by BOTH retrievers near the
+    top of each, to rank no lower than the runner-up in either source, and the
+    runner-up to be meaningfully weaker. Otherwise returns False and the caller
+    falls back to the pure ratio/gap gate. Agreement can only LIFT — the demotion
+    gates still apply.
+
+    ``vector_leg_keyless`` swaps the vector leg for the symbol leg. On an index
+    with no semantic vectors the vector leg is skipped outright, so ``_vec_rank``
+    is never written for any question and a fixed FTS+vector pair makes this
+    signal permanently unreachable — every keyless answer is then graded by the
+    pure ratio gate, which is exactly the gate this function exists because it
+    mis-reads. The symbol leg runs on every index and records ``_sym_rank``.
+
+    **The caller must pass the retrieval leg's own status, not infer it from
+    the hits.** By the time this runs, ``hits`` is capped to the top 5 out of a
+    much larger fused pool, so "no hit carries a ``_vec_rank``" is *not*
+    evidence the leg was skipped: a keyed index whose vector leg timed out,
+    errored, was scope-filtered, or was simply outranked by five FTS-and-symbol
+    hits presents identically. Substituting on that inference would fire exactly
+    when evidence is weakest and manufacture "high" confidence from it.
+
+    The symbol pair is held to a stricter rank than the vector pair. FTS and the
+    symbol leg are not independent — the wiki page FTS indexes contains the
+    public symbol table the symbol leg matches on — so their agreeing is closer
+    to one lexical match observed twice than to two retrievers concurring, and
+    the fusion beside this already prices that leg well below the others.
+    Requiring an exact rank-0 tie keeps the weaker signal from carrying the
+    stronger claim.
+
+    Note the consequence: at ``top_rank_max = 0`` the runner-up comparison below
+    can no longer reject anything, because two hits cannot share rank 0 within
+    one leg. The symbol pair therefore reduces exactly to "the top hit is #1 in
+    FTS and #1 in the symbol leg", which is the intended rule; the shared gap
+    check is retained for the vector pair, where it does constrain.
+    """
+    if len(hits) < 2:
+        return False
+    if vector_leg_keyless:
+        second_field = "_sym_rank"
+        top_rank_max = _SYMBOL_AGREEMENT_TOP_RANK_MAX
+    else:
+        second_field = "_vec_rank"
+        top_rank_max = _AGREEMENT_TOP_RANK_MAX
+    top = hits[0]
+    top_a = top.get("_fts_rank")
+    top_b = top.get(second_field)
+    # Top must be a consensus pick: found by both retrievers, near the top of
+    # each. A one-retriever top hit is exactly the ambiguous case we must NOT
+    # lift.
+    if top_a is None or top_b is None:
+        return False
+    if top_a > top_rank_max or top_b > top_rank_max:
+        return False
+    second = hits[1]
+    sec_a = second.get("_fts_rank")
+    sec_b = second.get(second_field)
+    # Runner-up found by only one retriever -> the consensus top clearly wins.
+    if sec_a is None or sec_b is None:
+        return True
+    # Runner-up found by both: the top must rank at least as high in BOTH
+    # sources (no source disagrees) and strictly ahead in at least one.
+    if top_a <= sec_a and top_b <= sec_b:
+        return (sec_a - top_a) >= _AGREEMENT_RANK_GAP or (
+            sec_b - top_b
+        ) >= _AGREEMENT_RANK_GAP
+    return False
+
+
+def _retrieval_quality(hits: list[dict], agreement_dominant: bool) -> str:
+    """Rate the retrieval, independently of the text it fed.
+
+    Kept as one function because the degraded path needs the same rating and must
+    not invent a second one. That path has no synthesised text to rate (its
+    ``confidence`` stays low, correctly), but it ran exactly the same retrieval,
+    and "high" has to mean the same thing to a keyless caller as to a keyed one
+    or the field is worth less than nothing.
+    """
+    ratio = _top_two_score_ratio(hits)
+    top_score = hits[0].get("score", 0.0) if hits else 0.0
+    dominant_grade = ratio >= _DOMINANCE_RATIO or agreement_dominant
+    if dominant_grade and top_score >= _HIGH_CONFIDENCE_SCORE_FLOOR:
+        return "high"
+    return "partial" if dominant_grade else "weak"
+
+
+def _is_question_named_body_cut_by_us(entry: dict, question_ids: set[str]) -> bool:
+    """Whether this body is the question's own symbol, cut because WE ran out of lines.
+
+    ``truncated`` alone is not trustworthy enough to demote on. The stale-bound
+    case this was written for is now fixed at its source — ``check_symbol_bounds``
+    clamps every bound to the live file — but the guard still earns its keep on
+    the ``source_excerpt`` fallback, where the served bytes come from the index
+    rather than from disk and the live length says nothing. Requiring the served
+    span to have reached the line cap says the cut was ours, which is the only
+    case where something was really withheld.
+    """
+    if not (entry.get("truncated") and entry.get("continuation")):
+        return False
+    if entry.get("name") not in question_ids:
+        return False
+    return entry["lines"][1] - entry["lines"][0] + 1 >= _INLINE_BODY_MAX_LINES
+
+
+def _is_enclosing_continuation(entry: dict, implicated: set[str]) -> bool:
+    """Whether this body simply continues past the cut, rather than losing a symbol.
+
+    A withheld entry carrying the served body's OWN name is the enclosing symbol
+    continuing past the cut, not something that never arrived. Calling that "not
+    served" is wrong about the payload directly above the note, and sends the
+    caller to get_symbol for a body they already hold most of. The accurate
+    pointer is the ``continuation`` the entry already carries.
+    """
+    name = entry.get("name")
+    if not (entry.get("continuation") and name in implicated):
+        return False
+    return any(s.get("name") == name for s in (entry.get("withheld_symbols") or []))
+
+
+class _Grade(NamedTuple):
+    """The confidence verdict, and every finding the notes are written from.
+
+    The gates do not just produce a label: each one that fires records WHAT
+    it objected to, and the payload builder turns that into the note and the
+    next action. Carrying the findings out beside the verdict is what keeps
+    the two in step.
+    """
+
+    confidence: str
+    hedged: bool
+    ratio: float
+    top_score: float
+    ungrounded_values: list[str]
+    frame_unsupported: list[str]
+    exclusivity_over_truncated: bool
+    withheld_implicated: list[str]
+    lookup_body_truncated: bool
+    named_body_cut: dict | None
+
+
+def _grade_answer(
+    *,
+    question: str,
+    question_ids: set[str],
+    answer_text: str,
+    hits: list[dict],
+    citations: list[str],
+    symbol_bodies: list[dict],
+    served_named_body: bool,
+    dominant: bool,
+    agreement_dominant: bool,
+) -> _Grade:
+    """Grade the synthesised answer through the gate cascade, in order.
+
+    One starting grade from retrieval dominance, then a run of gates that can
+    only demote it. Several are guarded on the answer still being at high, so
+    that one response cannot be pushed two levels for one problem. Read them
+    as a list of reasons not to trust the prose: the order is the order they
+    were added, and each comment says which failure it was built to catch.
+    """
+    # Compute confidence from the dominance ratio (top hit vs second hit).
+    # The dominance ratio is a more reliable separator than absolute BM25
+    # thresholds, which tend to label most retrievals "high" indiscriminately.
+    _ratio = _top_two_score_ratio(hits)
+    _top_score = hits[0].get("score", 0.0) if hits else 0.0
+    # Agreement lifts the RRF-compressed ratio: a consensus top hit (both
+    # retrievers rank it at/near #1) grades dominant even though its fused
+    # score barely outscores the runner-up. The score floor still applies, and
+    # is naturally cleared — a rank-0-in-both hit scores well above it.
+    _dominant_grade = _ratio >= _DOMINANCE_RATIO or agreement_dominant
+
+    # Strong answer-grounding can EARN "high" on a NON-dominant retrieval. RRF
+    # compresses sibling scores, so a rank-1 hit buried in a cluster of related
+    # files never "dominates" numerically and was capped at medium even when the
+    # synthesised answer is fully grounded in served source. Earn high when
+    # EITHER the question's named symbol body is served in-hand (tier-0 anchor),
+    # OR every distinctive mechanism term the answer names is grounded in the
+    # retrieval corpus AND a cited hit carries real symbol bodies. Conservative:
+    # a single ungrounded mechanism term disqualifies (that is the fabricated-
+    # mechanism signal), and the score floor still applies — this lifts
+    # non-dominant *well grounded* answers, never weakly-retrieved ones. The
+    # demotion gates below (hedge, value, claim-support) still pull an earned
+    # high back down.
+    earn_high = False
+    if _flag_on(_EARN_HIGH_GROUNDING_ENV) and _top_score >= _HIGH_CONFIDENCE_SCORE_FLOOR:
+        _cited = set(citations)
+        _cited_has_body = any(h.get("symbols") for h in hits if h.get("target_path") in _cited)
+        _fu, _fg = _frame_term_grounding(answer_text, question, hits)
+        grounding_strong = _cited_has_body and _fg >= 1 and not _fu
+        earn_high = served_named_body or grounding_strong
+
+    if (_dominant_grade or earn_high) and _top_score >= _HIGH_CONFIDENCE_SCORE_FLOOR:
+        confidence = "high"
+    elif _dominant_grade:
+        # Dominant but weak — the right file relative to its siblings, but
+        # the signal isn't strong enough to trust the synthesised answer
+        # without verification. Downgrade so the consumer Reads the source.
+        confidence = "medium"
+    else:
+        confidence = "medium"
+
+    # Second gate: downgrade when the LLM's own answer admits insufficiency.
+    # Retrieval dominance only tells us we indexed the right file; it does
+    # not mean the synthesized text is usable. Shipping a hedged answer with
+    # confidence="high" misleads the consumer AND drags the full retrieval
+    # payload through the conversation cache for no benefit.
+    hedged = _answer_is_hedged(answer_text)
+    if hedged:
+        # A hedge means the synthesised PROSE is weak — but when the exact
+        # symbol the question named is inlined in symbol_bodies (tier-0 anchor,
+        # full live body), the answer's ground truth is already in-hand. Labeling
+        # that "low" contradicts the payload and fires the "go Read" hint the
+        # body makes unnecessary, so the agent bails to Read when it never needed
+        # to. Hold such a response at medium; the note redirects the agent from
+        # the hedged prose to the served body.
+        confidence = "medium" if served_named_body else "low"
+
+    # Third gate — identifier-citation gate: when the question explicitly
+    # names identifiers (classes / methods / snake_case / CamelCase) and
+    # NONE of the top retrieval hits contain any of those identifiers as a
+    # hydrated symbol, retrieval may be pointing at plausible-but-wrong
+    # files (same module family, similar vocabulary). Downgrade high->medium
+    # so the consumer Reads the `fallback_targets`. Only applies when the
+    # question actually names identifiers — mechanism-descriptive questions
+    # (no symbol names) are unaffected.
+    if confidence == "high" and question_ids:
+        top_n = [h for h in hits[:_ENRICH_TOP_N_HITS] if h.get("symbols")]
+        has_match = any(s.get("_matched") for h in top_n for s in (h.get("symbols") or []))
+        if not has_match:
+            confidence = "medium"
+
+    # Fourth gate — value grounding: on value-shaped questions (default /
+    # threshold / limit / how many), every number the answer asserts must
+    # appear somewhere in the material retrieval actually contained. A
+    # number synthesis produced from thin air is a factual error delivered
+    # with authority — the single worst calibration failure, because the
+    # consumer was told not to verify. Cap at low and say why.
+    ungrounded_values: list[str] = []
+    if not hedged and _is_value_question(question):
+        ungrounded_values = _ungrounded_numbers(answer_text, hits)
+        if ungrounded_values:
+            confidence = "low"
+
+    # Fifth gate — citation-source gate: a high-confidence answer must cite
+    # at least one page that contributed actual source material (hydrated
+    # symbols with signatures/bodies), not just file summaries. Summary-only
+    # grounding is how plausible-but-wrong syntheses get through.
+    if confidence == "high":
+        cited = set(citations)
+        if not any(h.get("symbols") for h in hits if h.get("target_path") in cited):
+            confidence = "medium"
+
+    # Sixth gate — claim-support / frame grounding: a high-confidence answer
+    # must name its mechanism in terms the cited material actually contains. The
+    # dominance gate is generous on repo-internal questions (an anchored symbol +
+    # a dominant hit clear it), so a synthesis that conflates two mechanisms —
+    # right file, wrong reason/function — rides through at high confidence. The
+    # tell is a distinctive code-like term (a class / function / module the
+    # answer names AS the mechanism) that appears nowhere in everything retrieval
+    # showed. When such terms are not outweighed by grounded ones, downgrade
+    # high->medium so the consumer verifies instead of trusting.
+    #
+    # The original gate fired only on "why" questions, but the same failure
+    # occurs on "how" questions that name the mechanism in the ANSWER, not the
+    # question. Value questions have their own numeric gate above; naming/lookup
+    # questions legitimately just echo the named symbol, so they are excluded.
+    frame_unsupported: list[str] = []
+    _claim_scope = _is_why_question(question) or (
+        _flag_on(_CLAIM_SUPPORT_GATE_ENV) and _is_mechanism_question(question)
+    )
+    if confidence == "high" and not hedged and _claim_scope:
+        frame_unsupported, _grounded_terms = _frame_term_grounding(answer_text, question, hits)
+        if frame_unsupported and len(frame_unsupported) >= _grounded_terms:
+            confidence = "medium"
+        else:
+            frame_unsupported = []
+
+    # Seventh gate — completeness scope over truncated bodies: prose asserts an
+    # unqualified exclusivity claim ("entirely", "the sole", "the only") while a
+    # cited symbol body arrived truncated, so the answer asserts a global
+    # property from a sample it knows is incomplete.
+    # Guard: only fires at high (like gates 3 / 5 / 6) so it cannot stack with
+    # other downgrades and push a single response two levels for one problem.
+    exclusivity_over_truncated = False
+    if confidence == "high" and not hedged:
+        exclusivity_over_truncated = _has_unqualified_exclusivity_over_truncated(
+            answer_text, symbol_bodies
+        )
+        if exclusivity_over_truncated:
+            confidence = "medium"
+
+    # Eighth gate — a WITHHELD symbol the response depends on. The seventh gate
+    # needs an exclusivity token in the prose as well as truncation, and across
+    # repeated runs of the reference defect the token usually does not appear:
+    # the rest are equally incomplete, equally "high", and it stays silent. It is
+    # also inert by construction in no-LLM mode, where there is no prose to hold
+    # a token.
+    #
+    # This gate keys on the dependency instead. It fires when a symbol in the
+    # withheld range is named by the QUESTION (every mode) or referenced as code
+    # by the ANSWER (LLM mode). Truncation alone is deliberately NOT enough: a
+    # large minority of truncations withhold nothing the response leans on, and
+    # `high` is worth keeping when it is earned.
+    withheld_implicated = implicated_withheld_symbols(question, answer_text, symbol_bodies)
+    if confidence == "high" and withheld_implicated:
+        confidence = "medium"
+
+    # Ninth gate — the LOOKUP half of the eighth, and a routing hole rather
+    # than a threshold problem. Gate 8 asks whether the question names a
+    # WITHHELD symbol; on a bare-name lookup it provably never can, because the
+    # question names the symbol that was SERVED and the withheld names are that
+    # symbol's own inner members. The union path caps on truncation alone for
+    # exactly this shape, on the grounds that where the caller asked for a symbol
+    # the bodies ARE the answer. But a name with a single definition never
+    # reaches the union path — `_anchor_symbol_hits` short-circuits at
+    # `len(cands) == 1` before `homonyms["union"]` is built — so it lands here
+    # with the same shape and, until now, no cap. The dominance ratio is not the
+    # cause and is not touched.
+    #
+    # Deliberately narrow. It needs the question to read as a symbol lookup
+    # rather than prose, AND the truncated body to be the very symbol the
+    # question named. On a prose question the body is evidence for a claim rather
+    # than the answer itself, and truncation alone is a poor signal there, which
+    # is why gate 8 keeps the dependency test for that population. Kept as the
+    # entry rather than a flag so the note can quote the real served range and
+    # continuation instead of describing the cut in the abstract.
+    named_body_cut = next(
+        (b for b in symbol_bodies if _is_question_named_body_cut_by_us(b, question_ids)),
+        None,
+    )
+    lookup_body_truncated = (
+        confidence == "high"
+        and named_body_cut is not None
+        and is_symbol_lookup_question(question, question_ids)
+    )
+    if lookup_body_truncated:
+        confidence = "medium"
+
+    # Non-dominant ceiling: ambiguous retrieval is the calibration cost of
+    # always synthesizing — the answer may be right, but with no single dominant
+    # page it must never read "high" (cite without verifying). Cap at medium even
+    # if every gate passed. (A non-dominant retrieval already scores <high via
+    # the ratio, so this is usually a no-op; it is explicit so the
+    # always-synthesize contract — "answered, but verify" — is self-documenting.)
+    # Exception: an answer that EARNED high via strong grounding (named symbol
+    # body in-hand, or every mechanism term grounded in a cited body) is not
+    # "cite without verifying" — the source IS in the payload — so the
+    # non-dominance ceiling does not apply to it.
+    if not dominant and not earn_high and confidence == "high":
+        confidence = "medium"
+    return _Grade(
+        confidence=confidence,
+        hedged=hedged,
+        ratio=_ratio,
+        top_score=_top_score,
+        ungrounded_values=ungrounded_values,
+        frame_unsupported=frame_unsupported,
+        exclusivity_over_truncated=exclusivity_over_truncated,
+        withheld_implicated=withheld_implicated,
+        lookup_body_truncated=lookup_body_truncated,
+        named_body_cut=named_body_cut,
+    )
